refactor out zeppelin websocket listener
